packages/pip/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zwischen"
version = "0.1.0"
description = "AI-augmented security scanning for vibe coders. Zero-config secrets detection and vulnerability scanning."
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Conner Jordan", email = "connercharlesjordan@gmail.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
keywords = ["security", "scanner", "secrets", "gitleaks", "semgrep", "ai", "vulnerability", "sast"]
requires-python = ">=3.9"
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
]

[project.scripts]
zwischen = "zwischen.cli:main"

[project.urls]
Homepage = "https://github.com/cjordan223/Zwischen"
Repository = "https://github.com/cjordan223/Zwischen.git"
Issues = "https://github.com/cjordan223/Zwischen/issues"

[tool.setuptools.packages.find]
where = ["."]
